requirements.txt
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-core>=1.29.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0 